:necktie: Add bootstrap WS connections for account transactions and tokens queries
Bootstrap WS connections for re-fetching account transactions and tokens data when
