+++ src/AppBootstrapper.js
@@ -1,14 +1,17 @@
 import { useEffect } from 'react';
+import { useQueryClient } from '@tanstack/react-query';
 
 import apiClient from 'utilities/api/APIClient';
+import { GET_ACCOUNT_TRANSACTIONS_QUERY, GET_ACCOUNT_TOKENS_QUERY } from 'utilities/api/queries';
 import { useCurrentApplication } from 'modules/BlockchainApplication/hooks/useCurrentApplication';
 
 /**
- * Bootstrap the app by calling all previous business logic to load
- * the required data.
+ * Bootstrap the app by calling all previous business logic to load the required data.
  * @param {ReactNode} children - Components tree to provide the loaded data.
  */
 export default function AppBootstrapper({ children }) {
+  const queryClient = useQueryClient();
+
   const [currentApplication] = useCurrentApplication();
 
   // Bootstrap API client with current application.
@@ -17,6 +20,23 @@ export default function AppBootstrapper({ children }) {
       apiClient.create(currentApplication.serviceURLs[0]);
     }
   }, [currentApplication]);
+
+  // Bootstrap WS connections for re-fetching account transactions and tokens data when
+  // new and delete transactions events occur.
+  useEffect(() => {
+    function invalidateQuery() {
+      queryClient.invalidateQueries([GET_ACCOUNT_TRANSACTIONS_QUERY], { exact: false });
+      queryClient.invalidateQueries([GET_ACCOUNT_TOKENS_QUERY], { exact: false });
+    }
+
+    apiClient?.ws?.on('new.transactions', invalidateQuery);
+    apiClient?.ws?.on('delete.transactions', invalidateQuery);
+
+    return () => {
+      apiClient?.ws?.off('new.transactions');
+      apiClient?.ws?.off('delete.transactions');
+    };
+  }, [queryClient]);
 
   return children;
 }
